Moved RDS Decoder into its own class
Moved RDS Decoder into its own class, allowing size optimizations and
making the code more modular and readable. Now Si4735 deals exclusively
with talking to the chip, as it should be.
Optimized out frequency and RSQ inside Si4735: we always ask the chip
when returning those so why store them? Memory is expensive on a
microcontroller.