Disabled I2C and switched to bool
- the pointer corruption problems I've seen lately have been traced to
the Wire library, I've disabled it while an investigation is underway.
A refactor + facelift could be in order for Wire as well since, IMHO,
it was way too many lines (and files) of code for what it does (compare
with SPI).
- I do enjoy standards quite a lot, even if Arduino differs: switched
to bool, ditched boolean. Compiles and works just fine, code size
smaller by 5 bytes :-)